Adds site_reports back temporarily
This is so we can clone this repo from the Macbook Pro, because the Macbook Pro is not able to run the sitemap_gen.py, due to problems installing Reppy there. But link_checker should work, so I ran the site_report.py to generate the site map then I'll run the link_checker.py on Macbook pro.